requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
chromadb
langchain
langchain-community
requests
python-multipart
pymupdf
python-docx
redis
celery
python-dotenv
reportlab
pytest
websockets